q2_plaigerism_detector/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
sentence-transformers>=2.2.0
openai>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.17.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
python-dotenv>=1.0.0 